requirements.txt
# FastAPI and ASGI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Redis
redis==5.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and Settings
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Celery for async tasks
celery[redis]==5.4.0

# File handling
boto3==1.35.29
python-magic==0.4.27

# Excel import/export
openpyxl==3.1.5
pandas==2.2.3

# HTTP client for webhooks
httpx==0.27.2

# Google Calendar
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.147.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2
